requirements.txt
Flask==3.0.2
APScheduler==3.10.4
python-dotenv==1.0.1
google-api-python-client==2.130.0   # Drive
google-auth==2.30.0
requests==2.32.2                    # Slack
python-dateutil==2.9.0